+++ src/components/MenuBar/provider.tsx
@@ -1,0 +1,4 @@
+const pages = ["Área de Trabalho"];
+const settings = ["Perfil", "Conta", "Sair"];
+
+export { pages, settings };
