+++ src/components/MenuBar/styles.tsx
@@ -1,0 +1,12 @@
+import { makeStyles } from "@mui/styles";
+
+const useStyles = makeStyles({
+  appBar: { background: "rgba(255,255,255, 0.4)" },
+  avatarButton: { padding: "0" },
+  avatar: { background: "#E7970D" },
+  menu: { marginTop: "45px" },
+  lightButton: { marginLeft: "10px" },
+  box: { flexGrow: 0 },
+});
+
+export default useStyles;
